+++ hardpy/hardpy_panel/frontend/src/hardpy_test_view/SuiteList.tsx
@@ -88,7 +88,8 @@ interface Props extends WithTranslation {
   defaultClose: boolean;
   onTestsSelectionChange?: (selectedTests: string[]) => void;
   selectedTests?: string[];
-  selectionSupported?: boolean; 
+  selectionSupported?: boolean;
+  measurementDisplay?: boolean;
 }
 
 const SECONDS_TO_MILLISECONDS = 1000;
@@ -267,6 +268,7 @@ export class SuiteList extends React.Component<
         onTestsSelectionChange={this.props.onTestsSelectionChange}
         selectedTests={this.props.selectedTests}
         selectionSupported={this.props.selectionSupported}
+        measurementDisplay={this.props.measurementDisplay}
       />
     );
   }