+++ hardpy/hardpy_panel/frontend/src/hardpy_test_view/TestSuite.tsx
@@ -90,6 +90,18 @@ interface DialogBoxProps {
   html?: HTMLInfo;
 }
 
+interface Measurement {
+  name?: string;
+  result?: boolean;
+  type: string;
+  value: number | string;
+  unit?: string;
+  comparison_value?: number | string;
+  operation?: string;
+  lower_limit?: number;
+  upper_limit?: number;
+}
+
 /**
  * Interface representing a test case
  * @interface Case
@@ -110,6 +122,7 @@ interface Case {
   stop_time: number;
   assertion_msg: string | null;
   msg: string[] | null;
+  measurements: Measurement[];
   artifact: Record<string, unknown>;
   dialog_box: DialogBoxProps;
   chart?: ChartData;
@@ -179,6 +192,7 @@ type Props = {
   selectionSupported?: boolean;
   moduleTechName: string;
   selectedTests?: string[];
+  measurementDisplay?: boolean;
 } & WithTranslation;
 
 /**
@@ -818,10 +832,12 @@ export class TestSuite extends React.Component<Props, State> {
         <TestData
           assertion_msg={test.assertion_msg}
           msg={test.msg}
+          measurements={test.measurements} 
           testSuiteIndex={this.props.index}
           testCaseIndex={rowIndex}
           chart={test.chart}
           dataColumnWidth={this.state.dataColumnWidth}
+          measurementDisplay={this.props.measurementDisplay}
         />
       </div>,
       `data_${rowIndex}_${row_}`