+++ docs/changelog.md
@@ -2,6 +2,9 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add measurement display functionality to operator panel showing numeric and string measurements with name-value-unit formatting.
+  [[PR-223](https://github.com/everypinio/hardpy/pull/223)]
+
 ## 0.17.1
 
 * Fix issue where when using a large button, it overlaps the content.